moved issue to seperate controller and added launch darkly flag
