+++ src/main/java/uk/gov/hmcts/reform/prl/services/RequestUpdateCallbackService.java
@@ -95,8 +95,8 @@ public class RequestUpdateCallbackService {
                             ? PAYMENT_SUCCESS_CALLBACK : PAYMENT_FAILURE_CALLBACK
             );
 
-            solicitorEmailService.sendEmail(caseDetails);
-            caseWorkerEmailService.sendEmail(caseDetails);
+//            solicitorEmailService.sendEmail(caseDetails);
+//            caseWorkerEmailService.sendEmail(caseDetails);
 
         } else {
             log.error("Case id {} not present", serviceRequestUpdateDto.getCcdCaseNumber());
